requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
idna==3.11
pillow==12.0.0
psycopg==3.2.13
psycopg-binary==3.2.13
pymysql==1.1.2
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
